+++ app/src/main/java/com/whzl/mengbi/ui/fragment/main/FollowFragment.java
@@ -80,7 +80,7 @@ public class FollowFragment extends BaseFragment implements OnRefreshListener, O
     }
 
     public void getAnchorList(int pager) {
-        int userId = (int) SPUtils.get(getContext(), SpConfig.KEY_USER_ID, 0);
+        long userId = (long) SPUtils.get(getContext(), SpConfig.KEY_USER_ID, 0L);
         HashMap hashMap = new HashMap();
         hashMap.put("userId", userId);
         hashMap.put("pager", pager);
